% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_processing.R
\name{merge_barcodes}
\alias{merge_barcodes}
\title{Barcode collapsing tool for single cell RNA seq by split-pool barcoding}
\usage{
merge_barcodes(matrix, bc1_list, output = getwd(), bc1_processed = NULL)
}
\arguments{
\item{matrix}{a dgT type matrix with barcodes as column names, in the format "ATGCCTAA_AATCCGTC_AAACGATA" and genes as row names}

\item{bc1_list}{a table containing a 'barcode1' columm and a 'barcode2' column, filled with the two possible round 1 barcodes in each well}

\item{output}{output directory. Current directory is default}

\item{bc1_processed}{optional. If the function was run before, give the directory of the 'colnames_BC1_processed.RDS' output file}
}
\value{
A new dgT type matrix with updated column names and summed counts
}
\description{
Barcode collapsing tool for single cell RNA seq by split-pool barcoding
}
